Webserver/requirements.txt
django
djangorestframework
requests
numpy
pymongo
ultralytics